refactor: rename and DRY-up code
