--- packages/blockchain/src/index.ts
@@ -1,4 +1,0 @@
-export { Blockchain, Wallet } from './blockchain'
-export { EVM } from './evm'
-export { Mina } from './mina'
-export { HDPath, KeyConst, Network, PathLevelIndexes } from './types'
