+++ packages/key-generator/tsconfig.json
