+++ packages/key-generator/tsup.config.ts
